renamed namespace and assembly first iteration of improvement exception handling fixed uint32_t type on ft_idx

+++ vowpalwabbit/interactions.h
@@ -58,13 +58,13 @@ void eval_count_of_generated_ft(vw& all, example& ec, size_t& new_features_cnt, 
 // 2 template functions to pass T() proper argument (feature idx in regressor, or its coefficient)
 
 template <class R, void (*T)(R&, const float, float&)>
-inline void call_T( R& dat, weight* weight_vector, const size_t weight_mask, const float ft_weight, const size_t ft_idx)
+inline void call_T( R& dat, weight* weight_vector, const size_t weight_mask, const float ft_weight, const uint32_t ft_idx)
 {
     T(dat, ft_weight, weight_vector[ft_idx & weight_mask]);
 }
 
 template <class R, void (*T)(R&, float, uint32_t)>
-inline void call_T( R& dat, weight* /*weight_vector*/, const size_t /*weight_mask*/, const float ft_weight, const size_t ft_idx)
+inline void call_T( R& dat, weight* /*weight_vector*/, const size_t /*weight_mask*/, const float ft_weight, const uint32_t ft_idx)
 {
     T(dat, ft_weight, ft_idx);
 }
@@ -155,7 +155,7 @@ inline void generate_interactions(vw& all, example& ec, R& dat, v_array<feature_
 
                     for (; fst != fst_end; ++fst)
                     {
-                        const uint32_t halfhash = FNV_prime * (uint32_t)fst->weight_index;
+                        const uint32_t halfhash = FNV_prime * fst->weight_index;
                         call_audit<R ,audit_func>(dat, fst);
                         // next index differs for permutations and simple combinations
                         const feature_class* snd = (!same_namespace) ? features_data[snd_ns].begin :
